chore: moved UI back in component, added test todo
