+++ src/components/DynamicFormContainer/DynamicFormLayout.tsx
@@ -4,9 +4,9 @@ import React, { FunctionComponent, useState } from "react";
 import { Redirect } from "react-router";
 import { useConfigContext } from "../../common/context/config";
 import { useFormsContext } from "../../common/context/forms";
-import { Button } from "../../UI/Button";
-import { SvgIcon, SvgIconTrash, SvgIconXCircle } from "../../UI/SvgIcon";
-import { ToggleSwitch } from "../../UI/ToggleSwitch";
+import { Button } from "../UI/Button";
+import { SvgIcon, SvgIconTrash, SvgIconXCircle } from "../UI/SvgIcon";
+import { ToggleSwitch } from "../UI/ToggleSwitch";
 import { Container } from "../Container";
 import { DataFileButton } from "./DataFileButton";
 import { DeleteModal } from "./DeleteModal";
