+++ src/components/PublishContainer/PublishPage/PublishPage.tsx
@@ -5,9 +5,9 @@ import { useConfigContext } from "../../../common/context/config";
 import { useFormsContext } from "../../../common/context/forms";
 import { usePublishQueue } from "../../../common/hook/usePublishQueue";
 import { Config } from "../../../types";
-import { Button } from "../../../UI/Button";
-import { SvgIcon, SvgIconCheckCircle } from "../../../UI/SvgIcon";
-import { Title } from "../../../UI/Title";
+import { Button } from "../../UI/Button";
+import { SvgIcon, SvgIconCheckCircle } from "../../UI/SvgIcon";
+import { Title } from "../../UI/Title";
 import { Container } from "../../Container";
 import { NavigationBar } from "../../NavigationBar";
 import { ProgressBar } from "../../ProgressBar";
