+++ src/components/UI/SvgIcon/SvgIcon.stories.tsx
