+++ src/components/UI/SvgIcon/SvgIcon.tsx
