+++ src/components/UI/Title/Title.tsx
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 import tw from "twin.macro";
-import { mixin } from "../../styles";
+import { mixin } from "../../../styles";
 
 export const Title = styled.h1`
   ${mixin.fontRobotoBold}
